Fixed up some warnings
Removed the interthread messaging.

Have not yet tested. Will post a message to the mailing list once my external app compiles against it.

+++ CKLeopardSourceListTableColumn.m
@@ -10,7 +10,7 @@
 
 @implementation CKLeopardSourceListTableColumn
 
-- (id)dataCellForRow:(int)row
+- (id)dataCellForRow:(NSInteger)row
 {
 	if (row >= 0)
 	{
